+++ packages/runtime/utils/src/domApis.ts
@@ -1,3 +1,4 @@
+import { __ELEMENT_DELETE_TAG__ } from './garfish';
 import { makeMap } from './utils';
 
 export interface Text {
@@ -120,6 +121,7 @@ export class DOMApis {
     const parentNode = el && el.parentNode;
     if (parentNode) {
       parentNode.removeChild(el);
+      el[__ELEMENT_DELETE_TAG__] = true;
     }
   }