+++ website/sidebars.js
@@ -92,5 +92,6 @@ module.exports = {
   guide: listDocIds('guide'),
   api: listDocIds('api'),
   issues: listDocIds('issues'),
+  changelog: listDocIds('changelog'),
   community: listDocIds('community'),
 };