+++ packages/runtime/browser-vm/CHANGELOG.md
@@ -3,6 +3,24 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.8](https://github.com/bytedance/garfish/compare/0.1.7...0.1.8) (2021-09-07)
+
+### Features
+
+- modify default export ([d288bce](https://github.com/bytedance/garfish/commit/d288bce9c6bf754c9803edc40a094ccf9147088f))
+
+## [0.1.7](https://github.com/bytedance/garfish/compare/v0.1.5...0.1.7) (2021-09-07)
+
+### Features
+
+- add the has proxy for document ([aa69074](https://github.com/bytedance/garfish/commit/aa69074787e651a71f56885ebadacd495fd92c02))
+
+## [0.1.6](https://github.com/bytedance/garfish/compare/v0.1.5...0.1.6) (2021-09-06)
+
+### Features
+
+- add the has proxy for document ([aa69074](https://github.com/bytedance/garfish/commit/aa69074787e651a71f56885ebadacd495fd92c02))
+
 ## [0.0.36](https://github.com/bytedance/garfish/compare/v0.0.35...0.0.36) (2021-07-29)
 
 ### Bug Fixes